Split holmeslibrary into mmap.py and services.py.
